AuthenticationAuthorizationService/src/main/java/com/cms/service/UserDetailsServiceImpl.java
package com.cms.service;


import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import com.cms.model.Login;
import com.cms.repository.UserRepository;

@Service
public class UserDetailsServiceImpl implements UserDetailsService {
	
	UserRepository userRepository;

	@Override
	@Transactional
	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
		return null;
	}

}
